Advance ECommerce Store API/6. Third Party Login with OAuth/requirements.txt
Flask
Flask-RESTful
Flask-JWT-Extended
Flask-SQLAlchemy
Marshmallow
Flask-Marshmallow
Marshmallow-SQLAlchemy
Python-Dotenv
Flask-Uploads
Flask-Migrate
Flask-OAuthlib
psycopg2-binary